IITB/mysite/requirements.txt
gunicorn==19.9.0
Django==2.2.5
django-map-widgets==0.2.2
django-tinymce4-lite==1.7.5
django-heroku==0.3.1
whitenoise==4.1.3
dj-database-url==0.5.0
psycopg2==2.8.3